initial commit, start refactor
