+++ src/App.jsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useEffect } from 'react'
 import './App.css'
 import './index.css'
 
@@ -8,7 +8,7 @@ import PublishModal from './components/PublishModal/PublishModal'
 import LaunchModal from './components/LaunchModal/LaunchModal'
 import LaunchImageModal from './components/LaunchModal/LaunchImageModal'
 
-// redux imports
+import { GetCollectionsService } from './services/get-collections-service'
 import { useSelector } from 'react-redux'
 
 function App() {
@@ -21,15 +21,20 @@ function App() {
   const _showLaunchImageModal = useSelector(
     (state) => state.mainSlice.showLaunchImageModal
   )
+  useEffect(() => {
+    GetCollectionsService()
+  }, [])
 
   return (
-    <div className="App">
-      <PageHeader></PageHeader>
-      <Content></Content>
-      {_showPublishModal ? <PublishModal /> : null}
-      {_showLaunchModal ? <LaunchModal /> : null}
-      {_showLaunchImageModal ? <LaunchImageModal /> : null}
-    </div>
+    <React.StrictMode>
+      <div className="App">
+        <PageHeader></PageHeader>
+        <Content></Content>
+        {_showPublishModal ? <PublishModal /> : null}
+        {_showLaunchModal ? <LaunchModal /> : null}
+        {_showLaunchImageModal ? <LaunchImageModal /> : null}
+      </div>
+    </React.StrictMode>
   )
 }
