+++ src/utils/index.js
@@ -2,6 +2,5 @@ export { default as debounce } from './debounce'
 export { setupBounds } from './setupBounds'
 export { convertDate } from './datetime'
 export { convertDateForURL } from './datetime'
-export { setupArrayBbox } from './setupBbox'
 export { setupCommaSeparatedBbox } from './setupBbox'
 export { colorMap } from './colorMap'
